Added detail info

+++ src/TravelBundle/Controller/PlaceController.php
@@ -76,19 +76,21 @@ class PlaceController extends Controller
     }
 
     /**
-     * @Route("/view/{id}", name="place_view")
+     * @Route("/details/{id}", name="place_details")
      * @Security("is_granted('IS_AUTHENTICATED_FULLY')")
      * @param $id
      * @return \Symfony\Component\HttpFoundation\Response
      */
     public function viewAction($id){
         $place = $this->getDoctrine()->getRepository(Place::class)->findOneBy(['id' => $id]);
+        $reservations['past'] = $this->getDoctrine()->getRepository(Reservation::class)->findPast(['place' => $place]);
+        $reservations['recent'] = $this->getDoctrine()->getRepository(Reservation::class)->findRecent(['place' => $place]);
 
         if ($place === null){
             throw new \Exception('Undefined place');
         }
 
-        return $this->render('place/view.html.twig', ['place' => $place]);
+        return $this->render('front-end/place/details.html.twig', ['place' => $place, 'reservations' => $reservations]);
     }
 
     /**
@@ -142,9 +144,19 @@ class PlaceController extends Controller
     public function deleteAction($id){
         $place = $this->getDoctrine()->getRepository(Place::class)->findOneBy(['id' => $id]);
 
+        $reservations = $this->getDoctrine()->getRepository(Reservation::class)->findBy(['place' => $place]);
+
+        if (!empty($reservations)){
+            $this->addFlash('info', 'The place is booked. You cannot deleted it.');
+
+            return $this->redirectToRoute('place_details', ['id' => $id]);
+        }
+
         $em = $this->getDoctrine()->getManager();
         $em->remove($place);
         $em->flush();
+
+        $this->addFlash('info', 'The place was deleted successfully!');
 
         return $this->redirectToRoute('homepage');
     }
